+++ admin/admin_header.php
@@ -9,11 +9,11 @@
  * @license		http://www.gnu.org/licenses/old-licenses/gpl-2.0.html GNU General Public License (GPL)
  * @since		1.0
  * @author		Madfish (Simon Wilkinson) <simon@isengard.biz>
- * @package		quotes
+ * @package		testimonials
  * @version		$Id$
  */
 
 include_once "../../../include/cp_header.php";
 include_once ICMS_ROOT_PATH . "/modules/" . basename(dirname(dirname(__FILE__))) . "/include/common.php";
-if (!defined("QUOTES_ADMIN_URL")) define("QUOTES_ADMIN_URL", QUOTES_URL . "admin/");
-include_once QUOTES_ROOT_PATH . "include/requirements.php";
+if (!defined("TESTIMONIALS_ADMIN_URL")) define("TESTIMONIALS_ADMIN_URL", TESTIMONIALS_URL . "admin/");
+include_once TESTIMONIALS_ROOT_PATH . "include/requirements.php";