+++ admin/index.php
@@ -2,15 +2,15 @@
 /**
  * Admin index page of the module
  *
- * Including the quote page
+ * Including the testimonial page
  *
  * @copyright	Copyright Madfish (Simon Wilkinson) 2011
  * @license		http://www.gnu.org/licenses/old-licenses/gpl-2.0.html GNU General Public License (GPL)
  * @since		1.0
  * @author		Madfish (Simon Wilkinson) <simon@isengard.biz>
- * @package		quotes
+ * @package		testimonials
  * @version		$Id$
  */
 
-include 'quote.php';
+include 'testimonial.php';
 exit;