+++ icms_version.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Quotes version infomation
+ * Testimonials version infomation
  *
  * This file holds the configuration information of this module
  *
@@ -8,7 +8,7 @@
  * @license		http://www.gnu.org/licenses/old-licenses/gpl-2.0.html GNU General Public License (GPL)
  * @since		1.0
  * @author		Madfish (Simon Wilkinson) <simon@isengard.biz>
- * @package		quotes
+ * @package		testimonials
  * @version		$Id$
  */
 
@@ -18,9 +18,9 @@ defined("ICMS_ROOT_PATH") or die("ICMS root path not defined");
 $modversion = array(
 	"name"						=> "Testimonials",
 	"version"					=> "1.1",
-	"description"				=> _MI_QUOTES_MD_DESC,
+	"description"				=> _MI_TESTIMONIALS_MD_DESC,
 	"author"					=> "fiammybe (David Janssens)",
-	"credits"					=> "Thanks to Debianus for the quote rotator block and Madfish for the initial quote module",
+	"credits"					=> "Thanks to Debianus for the testimonial rotator block and Madfish for the initial testimonial module",
 	"help"						=> "",
 	"license"					=> "GNU General Public License (GPL)",
 	"official"					=> 0,
@@ -68,30 +68,30 @@ $modversion['people']['developers'][] = "fiammybe (David Janssens)";
 $modversion['manual']['wiki'][] = "<a href='https://github.com/fiammybe/testimonials/blob/master/docs/changelog.txt'>Documentation</a>";
 
 /** Database information */
-$modversion['object_items'][1] = 'quote';
+$modversion['object_items'][1] = 'testimonial';
 
 $modversion["tables"] = icms_getTablesArray($modversion['dirname'], $modversion['object_items']);
 
 /** Templates information */
 $modversion['templates'] = array(
-	array("file" => "quotes_admin_quote.html", "description" => "Quote admin index."),
-	array("file" => "quotes_requirements.html", "description" => "Display missing requirements."));
+	array("file" => "testimonials_admin_testimonial.html", "description" => "Testimonial admin index."),
+	array("file" => "testimonials_requirements.html", "description" => "Display missing requirements."));
 
 /** Blocks information */
 $modversion['blocks'][1] = array(
-	"file"						=> "quotes_random_quotes.php",
-	"name"						=> _MI_QUOTES_RANDOM_QUOTES,
-	"description"				=> _MI_QUOTES_RANDOM_QUOTESDSC,
-	"show_func"					=> "quotes_random_quotes_show",
-	"edit_func"					=> "quotes_random_quotes_edit",
+	"file"						=> "testimonials_random_testimonials.php",
+	"name"						=> _MI_TESTIMONIALS_RANDOM_TESTIMONIALS,
+	"description"				=> _MI_TESTIMONIALS_RANDOM_TESTIMONIALSDSC,
+	"show_func"					=> "testimonials_random_testimonials_show",
+	"edit_func"					=> "testimonials_random_testimonials_edit",
 	"options"					=> "0",
-	"template"					=> "quotes_random_quotes.html");
+	"template"					=> "testimonials_random_testimonials.html");
 
 $modversion['blocks'][] = array(
-	'file' => 'quotes_rotator_quotes.php',
-	'name' => _MI_QUOTES_ROTATION_QUOTES,
-	'description' => _MI_QUOTES_ROTATION_QUOTESDSC,
-	'show_func' => 'quotes_rotator_quotes_show',
-	'edit_func' => 'quotes_rotator_quotes_edit',
+	'file' => 'testimonials_rotator_testimonials.php',
+	'name' => _MI_TESTIMONIALS_ROTATION_TESTIMONIALS,
+	'description' => _MI_TESTIMONIALS_ROTATION_TESTIMONIALSDSC,
+	'show_func' => 'testimonials_rotator_testimonials_show',
+	'edit_func' => 'testimonials_rotator_testimonials_edit',
 	'options' => '5|0|300|6000|fade',
-	'template' => 'quotes_rotator_quotes.html');
+	'template' => 'testimonials_rotator_testimonials.html');