+++ include/requirements.php
@@ -6,7 +6,7 @@
  * @license		http://www.gnu.org/licenses/old-licenses/gpl-2.0.html GNU General Public License (GPL)
  * @since		1.0
  * @author		Madfish (Simon Wilkinson) <simon@isengard.biz>
- * @package		quotes
+ * @package		testimonials
  * @version		$Id$
  */
 
@@ -15,13 +15,13 @@ defined("ICMS_ROOT_PATH") or die("ICMS root path not defined");
 $failed_requirements = array();
 
 if (ICMS_VERSION_BUILD < 40) {
-	$failed_requirements[] = _AM_QUOTES_REQUIREMENTS_ICMS_BUILD;
+	$failed_requirements[] = _AM_TESTIMONIALS_REQUIREMENTS_ICMS_BUILD;
 }
 
 if (count($failed_requirements) > 0) {
 	icms_cp_header();
 	$icmsAdminTpl->assign("failed_requirements", $failed_requirements);
-	$icmsAdminTpl->display(QUOTES_ROOT_PATH . "templates/quotes_requirements.html");
+	$icmsAdminTpl->display(TESTIMONIALS_ROOT_PATH . "templates/testimonials_requirements.html");
 	icms_cp_footer();
 	exit;
 }