+++ language/german/common.php
@@ -6,20 +6,20 @@
  * @license		http://www.gnu.org/licenses/old-licenses/gpl-2.0.html GNU General Public License (GPL)
  * @since		1.0
  * @author		Madfish (Simon Wilkinson) <simon@isengard.biz>
- * @package		quotes
+ * @package		testimonials
  * @version		$Id$
  */
 
 defined("ICMS_ROOT_PATH") or die("ICMS root path not defined");
 
-// quote
-define("_CO_QUOTES_QUOTE_CREATOR", "Autor");
-define("_CO_QUOTES_QUOTE_CREATOR_DSC", "Person, von der dieses Zitat stammt.");
-define("_CO_QUOTES_QUOTE_DESCRIPTION", "Zitat");
-define("_CO_QUOTES_QUOTE_DESCRIPTION_DSC", "Hier ist das Zitat einzutragen.");
-define("_CO_QUOTES_QUOTE_DATE", "Datum");
-define("_CO_QUOTES_QUOTE_DATE_DSC", "Geben sie das Jahr an, indem das Zitat gemacht wurde.");
-define("_CO_QUOTES_QUOTE_ONLINE_STATUS", "Online?");
-define("_CO_QUOTES_QUOTE_ONLINE_STATUS_DSC", "Schalten Sie dieses Zitat online, oder offline.");
-define("_CO_QUOTES_QUOTE_ONLINE", "Online");
-define("_CO_QUOTES_QUOTE_OFFLINE", "Offline");
+// testimonial
+define("_CO_TESTIMONIALS_TESTIMONIAL_CREATOR", "Autor");
+define("_CO_TESTIMONIALS_TESTIMONIAL_CREATOR_DSC", "Person, von der dieses Zitat stammt.");
+define("_CO_TESTIMONIALS_TESTIMONIAL_DESCRIPTION", "Zitat");
+define("_CO_TESTIMONIALS_TESTIMONIAL_DESCRIPTION_DSC", "Hier ist das Zitat einzutragen.");
+define("_CO_TESTIMONIALS_TESTIMONIAL_DATE", "Datum");
+define("_CO_TESTIMONIALS_TESTIMONIAL_DATE_DSC", "Geben sie das Jahr an, indem das Zitat gemacht wurde.");
+define("_CO_TESTIMONIALS_TESTIMONIAL_ONLINE_STATUS", "Online?");
+define("_CO_TESTIMONIALS_TESTIMONIAL_ONLINE_STATUS_DSC", "Schalten Sie dieses Zitat online, oder offline.");
+define("_CO_TESTIMONIALS_TESTIMONIAL_ONLINE", "Online");
+define("_CO_TESTIMONIALS_TESTIMONIAL_OFFLINE", "Offline");