+++ language/german/main.php
@@ -6,11 +6,11 @@
  * @license		http://www.gnu.org/licenses/old-licenses/gpl-2.0.html GNU General Public License (GPL)
  * @since		1.0
  * @author		Madfish (Simon Wilkinson) <simon@isengard.biz>
- * @package		quotes
+ * @package		testimonials
  * @version		$Id$
  */
 
 defined("ICMS_ROOT_PATH") or die("ICMS root path not defined");
 
-define("_MD_QUOTES_ADMIN_PAGE", ":: Administration ::");
-define("_MD_QUOTES_ALL_QUOTES", "Alle Zitate");
+define("_MD_TESTIMONIALS_ADMIN_PAGE", ":: Administration ::");
+define("_MD_TESTIMONIALS_ALL_TESTIMONIALS", "Alle Zitate");