+++ CHANGELOG.md
@@ -4,7 +4,7 @@ A very simple module to manage and display testimonials about your product.
 ## Version 1.1 beta (30 Nov 2024)
 - Update module for PHP 7.4+
 ## Version 1.02 (14 Aug 2013)
-- Added new 'quote rotator' block (Debianus)
+- Added new 'testimonial rotator' block (Debianus)
 ## Version 1.01 (9 June 2012)
 - Fix slow admin index page load when gzip enabled on ICMS 1.3.2.
 ## Version 1.0